refactor: 모달 컴포넌트 분리 및 모달이름 수정
