+++ src/app/_components/molecules/ModalContent/editImageModal/EditImageModal.style.tsx
@@ -12,13 +12,17 @@ export const Content = {
     display: flex;
     justify-content: center;
     border-right: 1px solid rgb(219, 219, 219);
-
+    max-width: 1100px;
+    min-width: 35vw;
+    overflow: hidden;
     img {
-      width: 100%;
+      width: auto;
+      height: 100%;
     }
   `,
   Flex: styled.div`
     display: flex;
     flex: 1;
+    width: 100%;
   `,
 };
