+++ src/app/_components/molecules/ModalContent/editImageModal/EditImageModal.tsx
@@ -7,7 +7,6 @@ import useModalStore from "@/_stores/client/modalStore";
 import Image from "next/image";
 import IconButton from "@components/atoms/button/IconButton";
 import { MODAL_NAME } from "@/_constant/modal";
-import WritingBox from "../../writingBox/WritingBox";
 
 export default function EditImageModal() {
   const { modalName, selectedImage, setModal } = useModalStore();
@@ -26,7 +25,7 @@ export default function EditImageModal() {
       <ModalStyle.Header>
         <Content.Header>
           <IconButton awesomeIcon={faArrowLeft} color="gray" onClick={handleBackClick} />
-          {modalName === MODAL_NAME.WRITE_POST ? "새 게시물 만들기" : "미리보기"}
+          미리보기
           <BaseButton value="다음" color="#0095F6" onClick={handleNextClick} />
         </Content.Header>
       </ModalStyle.Header>
@@ -35,7 +34,6 @@ export default function EditImageModal() {
           <Content.ImgBox>
             {selectedImage && <Image src={selectedImage} width={300} height={400} alt="preview" />}
           </Content.ImgBox>
-          {modalName === MODAL_NAME.WRITE_POST && <WritingBox />}
         </Content.Flex>
       </ModalStyle.Body>
     </>
