+++ src/app/_constant/modal.ts
@@ -1,6 +1,6 @@
 export const MODAL_NAME = {
   POST_FEED: "새 게시물 만들기",
   EDIT_IMAGE: "이미지 자르기",
-  WRITE_POST: "게시물 만들기",
+  WRITE_POST: "게시글 작성하기",
   TEST: "TEST MODAL",
 };
