+++ src/app/_utils/utils.tsx
@@ -1,6 +1,7 @@
 import { MODAL_NAME } from "@/_constant/modal";
-import PostModal from "@components/molecules/ModalContent/postModal/PostModal";
-import EditImageModal from "@components/molecules/ModalContent/editModal/EditImageModal";
+import UploadModal from "@components/molecules/ModalContent/uploadModal/UploadModal";
+import EditImageModal from "@/_components/molecules/ModalContent/editImageModal/EditImageModal";
+import EditPostModal from "@components/molecules/ModalContent/editPostModal/EditPostModal";
 import TestModal from "@components/molecules/ModalContent/TestModal";
 
 export function getRandomBoolean() {
@@ -14,11 +15,11 @@ export function getRandomNumber(max: number) {
 export function getModal(modalName: string) {
   switch (modalName) {
     case MODAL_NAME.POST_FEED:
-      return <PostModal />;
+      return <UploadModal />;
     case MODAL_NAME.EDIT_IMAGE:
       return <EditImageModal />;
     case MODAL_NAME.WRITE_POST:
-      return <EditImageModal />;
+      return <EditPostModal />;
     case MODAL_NAME.TEST:
       return <TestModal />;
     default:
